ingester: Move test-only code to list_series_test.go
i noticed this while reviewing 12333 because it doesn't handle `storage.ErrNotFound` errors

this used to be used in prod ocde before 7986
